udev: change structure in /run/udev/links to improve efficiency
The patch "udev: lock symlink directories in link_update()" adds locking
in the link_update() code path which figures out the highest-prio target for
any symlink claimed by a current uevent. While this fixes the incorrectness
of the current code and avoids races, it may also cause a slowdown of udev
operation due to lock contention; in particular, for symlinks which are claimed
by a lot of different devices.

Typical real-world examples would be multipath setups with 16 paths per device
and a dm-multipath device with higher link priority. According to my research
up to this point, the code with locking handles this amount of devices quite
well. But there are also pathological cases, where a symlink is claimed by
hundreds ot thousands of devices in a system. Although this indicates a system
misconfiguration, udev should be able to handle it without grinding to a halt
due to lock contention. An example for such symlinks is
"/dev/disk/by-partlabel/primary", which is claimed by every partition in a GPT
partition table on certain distributions, due to some weirdness of past
versions of the parted program.

Currently udev handles symlink targets using files that look like this:

  /run/udev/links/${encoded_symlink_name}/${dev_id}

where dev_id is b$major:$minor or c$major:$minor.

When an event is processed, udev adds or removes these files according to event
type and scans the whole directory /run/udev/links/${encoded_symlink_name} to
find the entry with the highest priority. In the process, it needs to get an
udev device ref for every entry and derive the priority from libudev.
If there are many contenders for the symlink at hand, this procedure is
inefficient.

This patch changes the data structure as follows:

  /run/udev/links/${encoded_symlink_name}/L:$prio/${dev_id}

Thus entries are grouped by priority, and finding the highest-prio entry is
much quicker because we hardly ever have more than 2 or 3 prio levels for any
given symlink. In combination with the locking patch, the time during which the
lock is held can be drastically reduced.

I verified that this patch doesn't cause any test suite regressions.