units: Remove ordering of systemd-journal-flush before systemd-tmpfiles-setup
Commit 74055aa7 (journalctl: add new --flush command and make use of it
in systemd-journal-flush.service) adds the ordering to
`systemd-journal-flush.service` before `systemd-tmpfiles-setup.service`.

This holds up reaching of the target *sysinit*, as
`systemd-tmpfiles-setup.service` is ordered before that target. Also,
`systemd-update-utmp.service.in` is ordered before that target, and
after `systemd-tmpfiles-setup.service`.

As `journalctl --flush` takes around one second even on systems with an
SSD, this holds up the boot considerably. Therefore, remove that
ordering.

Reading the description of the switch `--flush` of `journalctl` it
should be fine to run it without having to order before anything else.