umask-util: implement RUN_WITH_UMASK using MPP_DECLARE
The implementation is simpler, doesn't require a boolean and should
yield fewer spurious warnings about variables not being initialized.

NOTE: One use in tmpfiles.c needed an extra set of braces, otherwise the
compiler warns about the code depending on the `else` clause binding to
the closest `if`. This is unfortunate, but to some extent expected:
https://www.chiark.greenend.org.uk/~sgtatham/mp/#S6