nspawn: change type of arg_settings_mask from SettingsMask to uint64_t
SettingsMask may be implemented as a 32 bit value. If so, some flags
for RLIMIT cannot be set. We usually use unsigned or uint64_t for bit
field variables. So, let's change the type of arg_settings_mask from
SettingsMask to uint64_t.
Also, this also makes SETTING_RLIMIT_FIRST be casted to uint64_t before
it is shifted, as it may also be implemented as a 32 bit value.

Fixes #9035.