login: protect FDSTOREREMOVE=1 against ENOMEM
Close #8567. I wrote that issue because I didn't like ignoring ENOMEM
for an sd_notify() call, which was itself intended to free memory.  It
felt like a potential deadlock.

In case Linux has deliberately been configured to not overcommit
memory, using a stack buffer will make this reliable.

By default Linux allows a certain amount of overcommit, so there will
be some remaining risk of OOM-kill and maybe an infinite restart loop.
A restart loop could impose a load that makes the system harder to
inspect.

Optimistically: the OOM could be easy to find in the logs.  Arguably
making it clearer what's happened to the system, than if we logged a
single "Out of memory" message and kept running.

Pessimistically: there's not much more I can do in logind.  OOM-kill can
strike at literally any point anyway, and the concern is more that
in general, systemd's time-based restart-loop protection could become
ineffective when system performance is degraded ("thrashing").  Maybe
there's room for some improvement there in future.

I didn't try to apply this trick to the general sd_notifyf() function.
It didn't seem very suited for stack allocations, as it is documented
in somewhat generic terms and without mentioning size limits.

AFAICT, resolving #8567 more comprehensively would require changes both
in the kernel and in the logind API.  (The suggested "steal master" drm
ioctl, and a requirement to pass your FD back to logind when when you
want to become master again.  Then logind simply doesn't need to put
drm FDs in the fdstore, and the need for FDSTOREREMOVE would go away).